Refactor spec directories
