+++ Gemfile
@@ -6,3 +6,9 @@ gem 'rspec'
 gem 'appraisal'
 gem 'rubocop'
 gem 'rubocop-md'
+gem 'dotenv', require: 'dotenv/load'
+gem 'mysql2'
+gem 'pg'
+gem 'activerecord-sqlserver-adapter'
+gem 'sqlite3'
+gem 'trilogy'
