+++ db/mysql/my.cnf
