+++ gemfiles/ar_7.0.gemfile
@@ -1,7 +1,20 @@
 # This file was generated by Appraisal
 
-source 'https://rubygems.org'
+source "https://rubygems.org"
 
-gem 'activerecord', '~> 7.0.0'
+gem "rspec"
+gem "appraisal"
+gem "rubocop"
+gem "rubocop-md"
+gem "dotenv", require: "dotenv/load"
+gem "mysql2"
+gem "pg"
+gem "activerecord-sqlserver-adapter"
+gem "sqlite3", "~> 1.4"
+gem "trilogy"
+gem "activerecord", "~> 7.0.0"
+gem "bigdecimal"
+gem "base64"
+gem "mutex_m"
 
-gemspec path: '../'
+gemspec path: "../"
