+++ gemfiles/ar_7.1.gemfile
@@ -6,6 +6,12 @@ gem "rspec"
 gem "appraisal"
 gem "rubocop"
 gem "rubocop-md"
+gem "dotenv", require: "dotenv/load"
+gem "mysql2"
+gem "pg"
+gem "activerecord-sqlserver-adapter"
+gem "sqlite3"
+gem "trilogy"
 gem "activerecord", "~> 7.1.0"
 
 gemspec path: "../"
