+++ gemfiles/ar_7.2.gemfile
@@ -6,6 +6,12 @@ gem "rspec"
 gem "appraisal"
 gem "rubocop"
 gem "rubocop-md"
+gem "dotenv", require: "dotenv/load"
+gem "mysql2"
+gem "pg"
+gem "activerecord-sqlserver-adapter"
+gem "sqlite3"
+gem "trilogy"
 gem "activerecord", "~> 7.2.0"
 
 gemspec path: "../"
