+++ spec/arproxy/plugin/query_logger.rb
