+++ spec/config_spec.rb
