+++ spec/integration/shared_examples/active_record_functions.rb
