+++ spec/integration/shared_examples/custom_proxies.rb
