+++ spec/integration/sqlserver_spec.rb
@@ -1,4 +1,4 @@
-require_relative 'spec_helper'
+require_relative '../spec_helper'
 
 context "SQLServer (AR#{ar_version})" do
   before(:all) do
